fix(canvas-note): scope note Cmd+S/Cmd+F to the focused note
Each mounted note attached its own window keydown listener, and the Cmd+S
handler never checked focus — so Cmd+S saved every open note at once.
Extract both global shortcuts into a useNoteKeyboard hook gated on
editor.isFocused, so a shortcut only acts on the note being edited.

As a side effect this trims useFileNodeEditor from 511 to 496 lines, back
under the file-size baseline.
